Changed types.

+++ src/trie_data/trie_data.c
@@ -18,8 +18,8 @@ trie_node *createTrieNode() {
 void addPrefixToTrie(trie_node *root, uint32_t base, char mask) {
     trie_node *current = root;
 
-    for (int i = MAX_MASK; i >= mask; i--) {
-        int bit = (base >> i) & 1;
+    for (uint8_t i = MAX_MASK; i >= mask; i--) {
+        uint8_t bit = (base >> i) & (uint32_t)1;
         if (current->children[bit] == NULL)
             current->children[bit] = createTrieNode();
         current = current->children[bit];
@@ -30,11 +30,11 @@ void addPrefixToTrie(trie_node *root, uint32_t base, char mask) {
 void deletePrefixFromTrie(trie_node *root, uint32_t base, char mask) {
     trie_node *current = root;
     trie_node *parent = NULL;
-    int bit;
+    uint8_t bit;
 
     // Przechodzenie do węzła zawierającego prefiks do usunięcia
-    for (int i = MAX_MASK; i >= mask; i--) {
-        bit = (base >> i) & 1;
+    for (uint8_t i = MAX_MASK; i >= mask; i--) {
+        bit = (base >> i) & (uint32_t)1;
         if (current->children[bit] == NULL)
             return; // Prefiks nie istnieje w drzewie
         parent = current;
@@ -56,8 +56,8 @@ char checkInTrie(trie_node *root, uint32_t ip) {
     trie_node *current = root;
     char maxMask = -1;
 
-    for (int i = MAX_MASK; i >= 0; i--) {
-        int bit = (ip >> i) & 1;
+    for (uint8_t i = MAX_MASK; i >= 0; i--) {
+        uint8_t bit = (ip >> i) & (uint32_t)1;
         if (current->children[bit] != NULL) {
             current = current->children[bit];
             if (current->mask != -1 && current->mask > maxMask)
@@ -74,7 +74,7 @@ void freeTrie(trie_node *root) {
     if (root == NULL)
         return;
 
-    for (int i = 0; i < 2; i++)
+    for (uint8_t i = 0; i < 2; i++)
         freeTrie(root->children[i]);
 
     free(root);
